+++ src/components/Views/Experiments/index.js
@@ -1,0 +1,9 @@
+import Experiments from './Home.js';
+import ExperimentsTable from './Table.js';
+import ExperimentsTableItem from './TableItem.js';
+
+export {
+  Experiments,
+  ExperimentsTable,
+  ExperimentsTableItem
+};
